refactor: add abstract suggestion class
